spotify/requirements.txt
Django==2.1.1
pytz==2019.3
mysqlclient==1.3.12
djangorestframework==3.10.3
requests==2.22.0
urllib3==1.25.7